pyproject.toml
[project]
name = "vector-databases"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langchain-classic",
    "langchain-chroma",
    "openai",
    "chromadb",
    "faiss-cpu",
    "numpy",
    "pandas",
    "python-dotenv",
    "pinecone",
    "sentence-transformers",
    "qdrant-client"
]
